usage ; lecture du pp.csv

+++ ENT/Identifiants_par_classe/identifiants_Kdecole.sh
@@ -6,13 +6,22 @@
 # Paramètres
 # 1 : nom de l'export Kd'école
 # 2 : type de profil voulu (Eleve, Tuteur, Enseignant, Personnel)
+# 3 : csv avec la liste des profs (format: adresse, classe)
+
+if [ $# -ne 3 ];then
+    echo "Usage : $0 Export_ENT.csv Profil Liste_prof.csv"
+    exit 1
+fi
+
 cat $1 | sort -t\; -k 6,6d -k 5,5d | awk -F';' -v profil=$2 '$4==profil {print $6","$5","$7","$8 > $3".csv"}'
 if [ -f classe.csv ]; then
     rm -f classe.csv
 fi
-# On tri alphabétiquement les fichiers générés
-#for cla in *.csv
-#do
-#	cat $cla | sort > tmp
-#    mv tmp $cla
-#done
+
+# Envoie par mail aux profs principaux
+for line in $3
+do
+    addr=echo $line | cut -d';' -f1
+    clas=echo $line | cut -d';' -f2
+    echo $addr - $clas
+done
